Text Editor: Add support for KTextEditor.
With these changes, KTextEditor can be used to edit text files like
python/macro files. These changes allow for the user to choose between
the built-in text editor and KTextEditor from the preferences editor.
This way, if KTextEditor has some breaking bug, then the user can use
the built-in text editor without having to recompile/download a
different version of FreeCAD. The preference setting used to choose the
preferred text editor is in a new tab in the Preferences dialog called
"Custom-Editor." This tab will be used to add more preferences settings
for KTextEditor in the future.

The implementation of KTextEditor introduced by this commit is really
simple, it can only open, edit, and close text files. However, it cannot
execute macros. Also, this implementation doesn't use FreeCAD commands
copy, paste, cut, undo, redo. Some of these commands will be used in the
future in order to have a good interface between KTextEditor and
FreeCAD.